Unit test update, the library test coverage is 88.38%

+++ IO/IKeySequence.cs
@@ -19,9 +19,9 @@ namespace Steganography.IO
 		/// <summary>
 		/// Get a sub set of the infinite sequence
 		/// </summary>
-		/// <param name="startAt">Index to start from</param>
+		/// <param name="index">Index to start from</param>
 		/// <param name="count">Size of the sub sequence</param>
 		/// <returns>Enumerable of bytes</returns>
-		IEnumerable<byte> GetSequence(int index, int count);
+		IEnumerable<byte> GetSequence(long index, int count);
 	}
 }
